Add files via upload
Run and Jump is a game in which you can collect points by collecting balls, and if the ball was missed, the points are reduced.

The player has 3 lives, which are reduced if he touches the saw. ⚙️

The saw needs to be jumped by pressing the space bar.

The game ends if the lives become equal to zero or the points fall below zero. 🔮

You can then restart the game without leaving the app! 🔥🔥🔥